apiserver: move patch tests to their own file
